+++ Source/Gui/Windows/MessageBox.h
@@ -33,7 +33,7 @@ public:
 			RGB565(000000), RGBTRANS, m_pszId);
 
 		rcClient.Deflate(2+12+4, 2+14+2, 2+12+4, 2+2+0);
-		int nExtentW = strlen(m_strMessage)*8;
+		int nExtentW = narrow_cast<int>(strlen(m_strMessage)*8);
 		int nExtentH = 14;
 
 		BIOS::LCD::Print( rcClient.Center().x - (nExtentW >> 1), rcClient.Center().y - (nExtentH >> 1 ),
